12.4+ Work-around for Zlib Compression KP
+ some serious README rewrite.

I was sorta late with this one I know but my machine is pretty basic non upgraded Penryn and can feel quite slow to the point you just don't wanna work on it haha.

Yet this is project to desperately keep my childhood machine alive.

Next on a list is personally hated, but must be done, EMULATED NVRAM, see ya soon in this madness.